refactor object helper and database access, start to add caching where appropriate

+++ object/core/model/ObjectTypeDaoMethods.java
@@ -6,7 +6,7 @@ public
 interface ObjectTypeDaoMethods {
 
 	ObjectTypeRec findById (
-			int id);
+			Long id);
 
 	ObjectTypeRec findByCode (
 			String code);
